requirements.txt
Flask==3.0.3
opencv-python-headless==4.9.0.80
numpy==1.24.3
pyserial==3.5
tflite-runtime==2.14.0
Werkzeug==3.0.1
Pillow==10.0.0